backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
supabase>=2.0.0
pyjwt>=2.8.0
cryptography>=41.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
